Promote shared modules to dirs
